feat: wip - finalize Tiki Crawler integration with job and schedule definitions, update dependencies, and enhance configuration
