+++ src/dagster_workspace/definitions.py
@@ -1,10 +1,12 @@
 import dagster as dg
 
-from assets.factory import build_dlt_pipelines
+from assets import build_dlt_pipelines
+from assets import defs
 from dbt_project import build_dbt_project
 
 
-defs = dg.Definitions.merge(
+main_defs = dg.Definitions.merge(
     build_dlt_pipelines(),
     build_dbt_project(),
+    defs,
 )
