+++ src/dagster_workspace/pyproject.toml
@@ -5,12 +5,12 @@ description = "Rainbow Dagster Workspace"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
-    "dagster==1.10.6",
-    "dagster-postgres>=0.26.6",
-    "dagster-obstore>=0.2.1",
-    "dagster-dlt>=0.26.6",
+    "dagster==1.10.9",
+    "dagster-postgres>=0.26.9",
+    "dagster-obstore>=0.2.3",
+    "dagster-dlt>=0.26.9",
     "dlt[clickhouse,postgres]>=1.9.0",
-    "dagster-dbt>=0.26.6",
+    "dagster-dbt>=0.26.9",
     "dbt-clickhouse>=1.8.9",
     "dbt-core>=1.8.9",
 ]
